clean up old comments and commented out code
